Added engine_server and engine_client as templates for further specializations.
git-svn-id: file:///home/pg512/share/svn/pg512/trunk@1470 0f804a50-a62f-0410-84e1-872c020a11f9

+++ honei/libvisual/graphrandom_TEST.cc
@@ -54,7 +54,7 @@ class EngineGraphTest :
 
             std::cout << "\nCalculate Position\n";
             
-            Engine::setTestCase(g, new Positions<Tag_, DataType_, GraphTag_>(g, (DataType_)1), _nodeCount * 0.1);
+            Engine::setTestCase(g, new Positions<Tag_, DataType_, GraphTag_>(g, (DataType_)1), int(_nodeCount * 0.1));
             
             
             char * c = "Test: Engine";
